working 1d fft kernel hw_emu and hw, small script to time fft on cpu

+++ FFT SW/fft.hxx
@@ -15,7 +15,7 @@
 
 using namespace std;
 
-typedef complex<double> cmpx_data_t;
+typedef complex<float> cmpx_data_t;
 
 
 /**
